+++ src/lsp/completions/tests/mod.rs
@@ -6,6 +6,7 @@ use {
 
 mod account_alias_tests;
 mod account_loader_tests;
+mod handler_call_return_tests;
 mod handler_member_tests;
 mod handler_method_tests;
 mod handler_transparent_tests;